fixup! many: refactor gpio helpers into separate sandbox/gpio package
